[INJIWEB-424] : vc verification before pdf download with error handling (#423)
* [INJIMOB-1679]: expose an api to verift the VC by consuming Vc-verifier library


* [INJIMOB-1679]: update multiple catch block structure


* [INJIMOB-1679]: update the vc-verify endpoint to expect VCCredentialResponse instead of String data


* [INJIMOB-1679]: refactor all controller tests


* [INJIMOB-1679]: rename ProofTypeNotFoundException to ProofTypeNotSupportedException


* [INJIWEB-424] : vc verification before pdf download


* [INJIWEB-424] : vc verification before pdf download


* [INJIWEB-424] : vc verification before pdf download with error handling


* [INJIWEB-424] : introduce error in vc verification


* [INJIWEB-424] : introduce error testing in vc verification


* [INJIWEB-424] : introduce error testing in vc verification


* [INJIWEB-424] : introduce error testing in vc verification


* [INJIWEB-424] : remove test ignore


* [INJIWEB-424] : modifying the utilities Bean


* [INJIWEB-424] : modifying the utilities Bean


* [INJIWEB-424] : modifying the utilities Bean


---------
