+++ src/main/java/io/mosip/mimoto/exception/PlatformErrorMessages.java
@@ -89,6 +89,7 @@ public enum PlatformErrorMessages {
     INVALID_CREDENTIAL_RESOURCE_URI_EXCEPTION(PlatformConstants.PREFIX + "040", "The Credential Resource is Invalid"),
     INVALID_CREDENTIAL_ISSUER_WELLKNOWN_RESPONSE_EXCEPTION(PlatformConstants.PREFIX + "041", "Invalid Wellknown from Issuer");
 
+
     /** The error message. */
     private final String errorMessage;
