+++ src/main/java/io/mosip/mimoto/service/impl/CredentialServiceImpl.java
@@ -19,6 +19,7 @@ import io.mosip.mimoto.dto.openid.presentation.PresentationDefinitionDTO;
 import io.mosip.mimoto.exception.ApiNotAccessibleException;
 import io.mosip.mimoto.exception.IdpException;
 import io.mosip.mimoto.exception.InvalidCredentialResourceException;
+import io.mosip.mimoto.exception.VCVerificationException;
 import io.mosip.mimoto.model.QRCodeType;
 import io.mosip.mimoto.service.CredentialService;
 import io.mosip.mimoto.service.IdpService;
@@ -28,8 +29,9 @@ import io.mosip.mimoto.util.LoggerUtil;
 import io.mosip.mimoto.util.RestApiClient;
 import io.mosip.mimoto.util.Utilities;
 import io.mosip.pixelpass.PixelPass;
+import io.mosip.vercred.CredentialsVerifier;
+import io.mosip.vercred.exception.*;
 import jakarta.annotation.PostConstruct;
-import lombok.NoArgsConstructor;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.Velocity;
 import org.jetbrains.annotations.NotNull;
@@ -49,6 +51,8 @@ import java.io.StringWriter;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
+
+import static io.mosip.mimoto.exception.ErrorConstants.*;
 
 @Service
 public class CredentialServiceImpl implements CredentialService {
@@ -95,9 +99,11 @@ public class CredentialServiceImpl implements CredentialService {
     PresentationServiceImpl presentationService;
 
     PixelPass pixelPass;
+    CredentialsVerifier credentialsVerifier;
     @PostConstruct
     public void init(){
         pixelPass = new PixelPass();
+        credentialsVerifier = new CredentialsVerifier();
     }
 
     @Override
@@ -118,8 +124,13 @@ public class CredentialServiceImpl implements CredentialService {
         CredentialsSupportedResponse credentialsSupportedResponse = issuerService.getIssuerWellknownForCredentialType(issuerId, credentialType);
         VCCredentialRequest vcCredentialRequest = generateVCCredentialRequest(issuerConfig, credentialIssuerWellKnownResponse,  credentialsSupportedResponse, response.getAccess_token());
         VCCredentialResponse vcCredentialResponse = downloadCredential(credentialIssuerWellKnownResponse.getCredentialEndPoint(), vcCredentialRequest, response.getAccess_token());
-        String dataShareUrl = dataShareService.storeDataInDataShare(objectMapper.writeValueAsString(vcCredentialResponse));
-        return generatePdfForVerifiableCredentials(vcCredentialResponse, issuerConfig, credentialsSupportedResponse, dataShareUrl);
+        boolean verificationStatus = vcCredentialResponse.getCredential().getProof().getType().equals("RsaSignature2018") ? verifyCredential(vcCredentialResponse) : true;
+        if(verificationStatus) {
+            String dataShareUrl = dataShareService.storeDataInDataShare(objectMapper.writeValueAsString(vcCredentialResponse));
+            return generatePdfForVerifiableCredentials(vcCredentialResponse, issuerConfig, credentialsSupportedResponse, dataShareUrl);
+        }
+        throw new VCVerificationException(SIGNATURE_VERIFICATION_EXCEPTION.getErrorCode(),
+                SIGNATURE_VERIFICATION_EXCEPTION.getErrorMessage());
     }
 
     public VCCredentialResponse downloadCredential(String credentialEndpoint, VCCredentialRequest vcCredentialRequest, String accessToken) throws InvalidCredentialResourceException {
@@ -149,6 +160,40 @@ public class CredentialServiceImpl implements CredentialService {
         LinkedHashMap<String, Object> displayProperties = loadDisplayPropertiesFromWellknown(vcCredentialResponse, credentialsSupportedResponse);
         Map<String, Object> data = getPdfResourceFromVcProperties(displayProperties, credentialsSupportedResponse,  vcCredentialResponse, issuerDTO, dataShareUrl);
         return renderVCInCredentialTemplate(data);
+    }
+
+    public Boolean verifyCredential(VCCredentialResponse vcCredentialResponse) throws VCVerificationException {
+        try {
+            String credentialString = objectMapper.writeValueAsString(vcCredentialResponse.getCredential());
+            return credentialsVerifier.verifyCredentials(credentialString);
+        } catch (ProofDocumentNotFoundException | ProofTypeNotSupportedException | SignatureVerificationException | UnknownException | JsonProcessingException |
+                 PublicKeyNotFoundException exception ) {
+            String errorCode = UNKNOWN_EXCEPTION.getErrorCode();
+            String errorMessage = UNKNOWN_EXCEPTION.getErrorMessage();
+            switch (exception.getClass().getSimpleName()) {
+                case "ProofDocumentNotFoundException" -> {
+                    errorCode = PROOF_DOCUMENT_NOT_FOUND_EXCEPTION.getErrorCode();
+                    errorMessage = PROOF_DOCUMENT_NOT_FOUND_EXCEPTION.getErrorMessage();
+                }
+                case "ProofTypeNotSupportedException" -> {
+                    errorCode = PROOF_TYPE_NOT_SUPPORTED_EXCEPTION.getErrorCode();
+                    errorMessage = PROOF_TYPE_NOT_SUPPORTED_EXCEPTION.getErrorMessage();
+                }
+                case "SignatureVerificationException" -> {
+                    errorCode = SIGNATURE_VERIFICATION_EXCEPTION.getErrorCode();
+                    errorMessage = SIGNATURE_VERIFICATION_EXCEPTION.getErrorMessage();
+                }
+                case "JsonProcessingException" -> {
+                    errorCode = JSON_PARSING_EXCEPTION.getErrorCode();
+                    errorMessage = JSON_PARSING_EXCEPTION.getErrorMessage();
+                }
+                case "PublicKeyNotFoundException" -> {
+                    errorCode = PUBLIC_KEY_NOT_FOUND_EXCEPTION.getErrorCode();
+                    errorMessage = PUBLIC_KEY_NOT_FOUND_EXCEPTION.getErrorMessage();
+                }
+            }
+            throw new VCVerificationException(errorCode, errorMessage);
+        }
     }
 
     @NotNull
